moved list of companies to main activity
